refactor: 불필요한 도메인로직 분리를 위한 커스텀훅 제거
