+++ src/components/InputField/CardNumberField.tsx
@@ -1,10 +1,17 @@
 import NumberInput from "../Input/NumberInput";
 import InputGroup from "./InputGroup";
-import useCardNumberField, {
-  type CardNumbers,
-  type CardBrand,
-} from "../../hooks/useCardNumberField";
+import { useFieldErrors } from "../../hooks/useFieldErrors";
+import { useRef } from "react";
+import { getCardNumberErrorMessage } from "../../utils/getCardNumberErrorMessage";
 
+export type CardNumbers = {
+  first: string;
+  second: string;
+  third: string;
+  fourth: string;
+};
+
+type CardBrand = "visa" | "master" | "diners" | "amex" | "unionpay" | null;
 interface Props {
   value: CardNumbers;
   cardBrand: CardBrand;
@@ -18,15 +25,44 @@ export default function CardNumberField({
   onComplete,
   cardBrand,
 }: Props) {
-  const {
-    inputErrors,
-    inputRefs,
-    errorMessage,
-    fourthMaxLength,
-    handleOnChange,
-    handleOnBlur,
-    setError,
-  } = useCardNumberField(value, cardBrand, onChange, onComplete);
+  const { inputErrors, setError, errorMessage } = useFieldErrors([
+    "first",
+    "second",
+    "third",
+    "fourth",
+  ]);
+
+  const inputRefs = useRef<Record<string, HTMLInputElement | null>>({});
+
+  const fourthMaxLength =
+    cardBrand === "diners" ? 2 : cardBrand === "amex" ? 3 : 4;
+
+  const handleOnChange = (cardKey: string) => (newValue: string) => {
+    setError(cardKey)(null);
+    const newCardNumbers = { ...value, [cardKey]: newValue };
+    onChange(newCardNumbers);
+
+    onComplete(getCardNumberErrorMessage(newCardNumbers, cardBrand) === null);
+
+    const maxLen = cardKey === "fourth" ? fourthMaxLength : 4;
+    const nextKey: Record<string, string> = {
+      first: "second",
+      second: "third",
+      third: "fourth",
+    };
+
+    if (newValue.length === maxLen && nextKey[cardKey]) {
+      inputRefs.current[nextKey[cardKey]]?.focus();
+    }
+  };
+
+  const handleOnBlur =
+    (cardKey: string) => (e: React.FocusEvent<HTMLInputElement>) => {
+      const currentValues = { ...value, [cardKey]: e.target.value };
+      const result = getCardNumberErrorMessage(currentValues, cardBrand);
+      const message = result && result.key === cardKey ? result.message : null;
+      setError(cardKey)(message);
+    };
 
   return (
     <InputGroup errorMessage={errorMessage}>
