+++ src/components/InputField/CvcNumberField.tsx
@@ -1,6 +1,8 @@
 import NumberInput from "../Input/NumberInput";
 import InputGroup from "./InputGroup";
-import useCvcNumberField from "../../hooks/useCvcNumberField";
+
+import { useState } from "react";
+import { getCvcNumberErrorMessage } from "../../utils/getCvcNumberErrorMessage";
 
 interface Props {
   onChange: (value: string) => void;
@@ -9,8 +11,18 @@ interface Props {
 }
 
 export default function CvcNumberField({ onChange, value, onComplete }: Props) {
-  const { inputError, setInputError, handleOnChange, handleOnBlur } =
-    useCvcNumberField(onChange, value, onComplete);
+  const [inputError, setInputError] = useState<string | null>(null);
+
+  const handleOnChange = (newValue: string) => {
+    setInputError(null);
+    onChange(newValue);
+    onComplete(getCvcNumberErrorMessage(newValue) === null);
+  };
+
+  const handleOnBlur = () => {
+    if (value !== "") setInputError(getCvcNumberErrorMessage(value));
+  };
+
   return (
     <InputGroup errorMessage={inputError}>
       <NumberInput
