+++ src/components/InputField/ExpNumberField.tsx
@@ -1,9 +1,11 @@
 import NumberInput from "../Input/NumberInput";
 import InputGroup from "./InputGroup";
-import useExpNumberField, {
-  type ExpNumber,
-} from "../../hooks/useExpNumberField";
 
+import { useRef } from "react";
+import { useFieldErrors } from "../../hooks/useFieldErrors";
+import { getExpNumberErrorMessage } from "../../utils/getExpNumberErrorMessage";
+
+export type ExpNumber = { mm: string; yy: string };
 interface Props {
   onChange: (value: ExpNumber) => void;
   value: ExpNumber;
@@ -11,14 +13,34 @@ interface Props {
 }
 
 export default function ExpNumberField({ onChange, value, onComplete }: Props) {
-  const {
-    inputRefs,
-    inputErrors,
-    setError,
-    errorMessage,
-    handleOnChange,
-    handleOnBlur,
-  } = useExpNumberField(onChange, value, onComplete);
+  const { inputErrors, setError, errorMessage } = useFieldErrors(["mm", "yy"]);
+
+  const inputRefs = useRef<Record<string, HTMLInputElement | null>>({});
+
+  const handleOnChange = (expKey: string) => (newValue: string) => {
+    setError(expKey)(null);
+    const newExpNumbers = { ...value, [expKey]: newValue };
+    onChange(newExpNumbers);
+    onComplete(getExpNumberErrorMessage(newExpNumbers) === null);
+
+    if (expKey === "mm" && newValue.length === 2) {
+      const result = getExpNumberErrorMessage({ mm: newValue, yy: value.yy });
+      if (result === null || result.key !== "mm") {
+        if (value.yy === "") inputRefs.current["yy"]?.focus();
+      } else setError("mm")(result?.message);
+    }
+  };
+
+  const handleOnBlur =
+    (expKey: string) => (e: React.FocusEvent<HTMLInputElement>) => {
+      const result = getExpNumberErrorMessage({
+        ...value,
+        [expKey]: e.target.value,
+      });
+      const message = result && result.key === expKey ? result.message : null;
+      setError(expKey)(message);
+    };
+
   return (
     <InputGroup errorMessage={errorMessage}>
       {Object.entries(value).map(([expKey, expValue]) => (
